Move Time/DateRange into util
