+++ src/main/java/uk/co/ramp/covid/simulation/covid/Covid.java
@@ -5,7 +5,7 @@
 package uk.co.ramp.covid.simulation.covid;
 
 import org.apache.commons.math3.random.RandomDataGenerator;
-import uk.co.ramp.covid.simulation.Time;
+import uk.co.ramp.covid.simulation.util.Time;
 import uk.co.ramp.covid.simulation.parameters.CovidParameters;
 import uk.co.ramp.covid.simulation.population.*;
 import uk.co.ramp.covid.simulation.util.Probability;
