+++ src/main/java/uk/co/ramp/covid/simulation/output/NetworkGenerator.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVPrinter;
 
-import uk.co.ramp.covid.simulation.Time;
+import uk.co.ramp.covid.simulation.util.Time;
 import uk.co.ramp.covid.simulation.place.Place;
 import uk.co.ramp.covid.simulation.population.Person;
