+++ src/main/java/uk/co/ramp/covid/simulation/place/Restaurant.java
@@ -1,15 +1,13 @@
 package uk.co.ramp.covid.simulation.place;
 
 import uk.co.ramp.covid.simulation.output.DailyStats;
-import uk.co.ramp.covid.simulation.Time;
+import uk.co.ramp.covid.simulation.util.Time;
 import uk.co.ramp.covid.simulation.population.Person;
 import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
 import uk.co.ramp.covid.simulation.population.Places;
 import uk.co.ramp.covid.simulation.population.Shifts;
 import uk.co.ramp.covid.simulation.util.RNG;
 import uk.co.ramp.covid.simulation.util.RoundRobinAllocator;
-
-import java.util.Iterator;
 
 public class Restaurant extends CommunalPlace {
