+++ src/main/java/uk/co/ramp/covid/simulation/util/Time.java
@@ -1,4 +1,4 @@
-package uk.co.ramp.covid.simulation;
+package uk.co.ramp.covid.simulation.util;
 
 import java.util.Objects;
