+++ src/test/java/uk/co/ramp/covid/simulation/TimeTest.java
@@ -1,6 +1,7 @@
 package uk.co.ramp.covid.simulation;
 
 import org.junit.Test;
+import uk.co.ramp.covid.simulation.util.Time;
 
 import static org.junit.Assert.*;
