+++ src/test/java/uk/co/ramp/covid/simulation/place/CommunalPlaceTest.java
@@ -3,7 +3,7 @@ package uk.co.ramp.covid.simulation.place;
 import org.junit.Before;
 import org.junit.Test;
 import uk.co.ramp.covid.simulation.output.DailyStats;
-import uk.co.ramp.covid.simulation.Time;
+import uk.co.ramp.covid.simulation.util.Time;
 import uk.co.ramp.covid.simulation.population.ImpossibleWorkerDistributionException;
 import uk.co.ramp.covid.simulation.population.Population;
 import uk.co.ramp.covid.simulation.testutil.PopulationGenerator;
