+++ src/test/java/uk/co/ramp/covid/simulation/population/PersonTest.java
@@ -1,7 +1,7 @@
 package uk.co.ramp.covid.simulation.population;
 
 import org.junit.Test;
-import uk.co.ramp.covid.simulation.Time;
+import uk.co.ramp.covid.simulation.util.Time;
 import uk.co.ramp.covid.simulation.testutil.SimulationTest;
 
 import static org.junit.Assert.*;
