+++ src/test/java/uk/co/ramp/covid/simulation/population/PopulationTest.java
@@ -1,9 +1,7 @@
 package uk.co.ramp.covid.simulation.population;
 
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
-import uk.co.ramp.covid.simulation.Time;
 import uk.co.ramp.covid.simulation.output.DailyStats;
 import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
 import uk.co.ramp.covid.simulation.place.*;
